+++ History.md
@@ -2,6 +2,7 @@
 ==================
   * Synchronize threads saving JSON to the same report directory (https://github.com/tntim96/JSCover/pull/40).
   * Fix logger NPE in InstrumentingRequestHandler
+  * Add Cobertura XML output (https://github.com/tntim96/JSCover/issues/39)
 
 0.2.1 / 2013-01-20
 ==================